Changes to be committed: 	modified:   controller/Controller.java 	modified:   intergration/CustomerPaymentDTO.java 	deleted:    model/Change.java 	modified:   model/Sale.java 	modified:   model/SaleInfo.java

+++ model/Sale.java
@@ -6,9 +6,9 @@ public class Sale {
 
 	private ItemDescriptionDTO soldItems;
 
-	private int totalPrice;
+	private double totalPrice;
 
-	private int totalVAT;
+	private double totalVAT;
 
 	private boolean itemFound;
